code review battery

+++ packages/notest-backend/src/battery/battery.controller.ts
@@ -1,11 +1,12 @@
 import { Body, Controller, Get, Post, Query } from '@nestjs/common';
 import { BatteryService } from './battery.service';
-import { UserId, UserIdIfHasToken } from '../shared/token.decorator';
+import { UserId } from '../shared/token.decorator';
 import { NTBattery } from '@notest/common';
 
 @Controller('battery')
 export class BatteryController {
   constructor(private batteryService: BatteryService) {}
+
   @Get('find-by-id')
   async getBatteryById(@Query('id') id: string) {
     return await this.batteryService.findById(id);
@@ -15,25 +16,24 @@ export class BatteryController {
   async getBatteryByUserid(@UserId() userId: number) {
     return await this.batteryService.findByField('userid', userId);
   }
-  @Post('create-battery')
-  async createBattery(@Body('battery') battery: NTBattery, @UserIdIfHasToken() userid) {
-    battery.userid = userid;
-    return await this.batteryService.create(battery);
+
+  @Post('save-battery')
+  async saveBattery(@Body('battery') battery: NTBattery) {
+    if (!battery.nt_batteryid) {
+      return await this.batteryService.create(battery);
+    } else {
+      await this.batteryService.update(battery);
+      if (battery.active) {
+        this.batteryService.updateCronJob(battery);
+      } else {
+        this.batteryService.deleteCronJob(battery.nt_batteryid);
+      }
+    }
   }
+
   @Get('delete-battery')
   async deleteBattery(@Query('battery_id') batteryId: string) {
     await this.batteryService.deleteCronJob(batteryId);
     await this.batteryService.delete(batteryId);
-  }
-  @Post('update-battery')
-  async updateBattery(@Body('battery') battery: NTBattery) {
-    console.log(battery);
-    await this.batteryService.update(battery);
-    if (battery.active) {
-      await this.batteryService.deleteCronJob(battery.nt_batteryid);
-      await this.batteryService.createCronJob(battery);
-    } else {
-      await this.batteryService.deleteCronJob(battery.nt_batteryid);
-    }
   }
 }
